testing: align CI gates and cleanup with the testing strategy
