+++ package.json
@@ -48,6 +48,9 @@
     "test:e2e:reduced": "node scripts/run-e2e.mjs --profile reduced",
     "test:scripts": "node scripts/run-tests.mjs --scope scripts",
     "test:packages": "node scripts/run-tests.mjs --scope packages",
+    "test:release-critical": "node scripts/testing-slices.mjs run release-critical-packages",
+    "test:release-critical:terminal": "node scripts/testing-slices.mjs run release-critical-terminal",
+    "test:progress": "node scripts/testing-slices.mjs summary",
     "hsr:demo:widget": "node scripts/hsr/widget-app.mjs",
     "hsr:demo:router": "node scripts/hsr/router-app.mjs",
     "hsr:record:widget": "node scripts/record-hsr-gif.mjs --mode widget",
